requirements.txt
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
sentence-transformers
python-dotenv
